pyproject.toml
[project]
name = "heysalad-platform"
version = "1.0.0"
description = "HeySalad Platform Backend - WMS, AI, Payments, Communication"
requires-python = ">=3.12"
dependencies = [
    # Core Framework
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
    
    # Database
    "supabase>=2.0.0",
    
    # AI Providers
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.3.0",
    
    # Payment
    "stripe>=7.0.0",
    
    # Communication
    "sendgrid>=6.11.0",
    "twilio>=8.10.0",
    
    # MCP Integration
    "fastmcp>=2.13.0.2",
    "mcp>=1.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.1.0",
]
